Separate search and prediction
In b788810e59, we integrated prediction into the pipeline step which also
was responsible for finding portfolios. However, different search
algorithms return different numbers of portfolios, which made prediction
tasks imbalanced in size (we need to make predictions independently for
each portfolio).
As a solution, we separate search and prediction. This also has the
advantage that one can easily comment out the prediction part or replace
the search part with loading existing search results if necessary. We
separate the results files for seach and prediction, but provide ids for
merging (though currently they same the same size and order, so can also
just append columns).