Remove tolerance from MIP search
When inspecting the results of the first experimental run, we noted that
the solution found by (supposedly exact) MIP search was sometimes worse
than that of (heuristic) beam search. In the worst case (for a portfolio
with k=39 and PAR 2 scoring), this resulted in a PAR2 increase of the
found portfolio by 0.0044 % - not much, but still annyoing.
We found this was the result of some tolerance in the configuration of the
optimizer, causing python-mip to stop search if the current optimum was
only a certain fraction away from a lower bound.
We disable this tolerance now.