Fix bug in storage of search function
Another fix for the apparently quite buggy commit 69177dedda. Since this
commit, the function object is stored for the search function instead of
the function name as string. Though the differene is not big, we revert to
the previous storage mode.
Also, we slightly simplify storage of search arguments, which was fixed in
4b877e931f.