Include results of 2021-10-04 experimental run
Only changes in the prediction pipeline (removal of gate recognition
features, introduction of xgboost), while search results remain same.
Change MCC plot to include xgboost and compare 2020 vs. 2021 as facets
(before, it only contained-random forest results and used the year for
coloring). Due to the changes in the feature set, the other two prediction
plots as well as the feature-importance analysis change slightly.