Adapt paper to LIPIcs template
Consider instructions and sample article from the template located at
https://submission.dagstuhl.de/styles/download/lipics/authors
(version 3.1.2 / instructions from 2021–06–26).

- revert some prior adaptations to LNCS template (32ad0242d6, 0ebc5ded5e)
- adapt preamble of document
- comment out packages preloaded by template
- move subsections in intro by one level (as formatting different now)
- replace $$ by \[ and \]
- switch to prescribed bib style

Apart from the preamble stuff, our document already was quite compliant to
the new template. We ignore the following recommendations for now:

- figures should be placed at top or bottom of page (we use htb)
- journal names should not be abbreviated (we consistently abbreviate)

Also, we leave the plots as-is, as they fit nicely into the current
template. The previous template already was single column, but had larger
margins. Thus, the font size of the figures is a bit larger than of the
main text now, but that does not hurt. Further, the font type of the main
text has only changed slightly (from Computer Modern to Latin Modern).