Try out regression-based solver selection
Proposed by a SAT reviewer: Rather than directly predicting the fastest
solver from the portfolio (which is a classification problem), predict the
runtime of each solver separately (which is a regression problem) and pick
the solver with the lowest predicted runtime. Unfortunately, this procedure
is rather slow (simply because we train more prediction models than before)
and the prediction performance is worse than for classification (maybe
because predicting runtimes is harder than recommending solver, and/or
the errors from the individual runtime prediction models build up to a high
total error).