Include kNN and MLP
Well, at least sort of. Both were proposed by a SAT reviewer, but did not
improve prediction performance. As they are fast to train (the MLP is only
scikit-learn's default one, a one-hidden-layer MLP), we include them
in the main experimental pipeline anyway, but prune them from evaluation
early and only spend a few sentences on them in the text.