Measure individual training + prediction time
Before, we only measured the total prediction time for one portfolio search
result, which involved multiple trainings + predictions, as the search
might yield multiple portfolios and we train multiple models for each
portfolio. Now, we store the time for each training + train/test prediction
separately.
However, this is only for diagnostic purposes, as the evaluation for the
paper does not use the prediction times anyway.