Parallelize over folds
Most of the pipeline' runtime is not spent on portfolio search, but for
training prediction models for 1000s of portfolios. Before, parallelization
was over datasets and search settings. However, the different search
settings yield a highly imbalanced number of portfolios: E.g., exact search
always yields one portfolio per k and dataset, while random search yield
1000 portfolios per k and dataset. To improve load balacing and also create
a bit more action on the console (the progress bar might not be updated for
hours), we additionally parallelize over folds now. As we only do 5-fold
cross-validation, this might not lead to huge improvements, but it won't
hurt either.