Slightly adapt plots' font type and size
Not done when switching to LIPIcs template in 74380f82b2, as the font type
was very similar to the previous template and the font size also fitted
really good to the text's / captions' size. We make very slight adaptations
in the code now (will update the plot files when the final pipeline run is
finished).