Remove MLP and KNN from main experiments
Introduced in fff6770b34, as we tried them out in preliminary experiments.
However, MLPs are really slow on the server (seem to be much faster
locally) and both types of models show significantly worse prediction
performance than the tree-based models, so we cut these approaches from
the main pipeline. We will add the MLP/kNN code to a separate branch.