Re-introduce search id
We removed the id identifying portfolio-search runs in commit 0250186afd,
because we thought it was unnecessary. After running the experimental
pipeline, we found out it actually had a purpose ... Having only problem
(= dataset) name, fold id, and solution (= portfolio) id is not sufficient
to join prediction results to search results, as these three attributes do
not uniquely identify individual portfolios (rows in the search results).
For example, multiple search runs yield an portfolio with solution_id = 0.
Thus, we re-introduce an id denoting the combination of dataset, fold, and
search settings (i.e., the tasks over which we parallelize).
Also, as beam search with w=100 yields more portfolios than random search
with w=1000 (for the 48 and 48 solvers in our datasets), we re-shuffle the
order of the search runs in the pipeline (to put heavier tasks to the
front).